requirements.txt
requests
textblob
pandas
python-dotenv
